Removed model file.
